requirements.txt
fastapi[standard]
pyyaml
